+++ src/main/java/nextstep/subway/station/infra/JpaStationRepository.java
@@ -1,5 +1,7 @@
-package nextstep.subway.station.domain;
+package nextstep.subway.station.infra;
 
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaStationRepository extends StationRepository, JpaRepository<Station, Long> {
