+++ src/test/java/nextstep/subway/path/application/PathServiceTest.java
@@ -1,10 +1,10 @@
 package nextstep.subway.path.application;
 
 import nextstep.subway.line.domain.LineRepository;
-import nextstep.subway.path.domain.InMemoryLineRepository;
+import nextstep.subway.line.infra.InMemoryLineRepository;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.application.StationService;
-import nextstep.subway.station.domain.InMemoryStationRepository;
+import nextstep.subway.station.infra.InMemoryStationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
